+++ lib/ftcircus/client.py
@@ -90,10 +90,16 @@ class CircusMgr(object):
                 "stdout_stream": {
                     "class": "FileStream",
                     "filename": u"%s/%s-stdout.log" % (logpath, name),
+                    "max_bytes": 2097152,
+                    # Removing backup logs for now to reduce clutter
+                    # "backup_count": 1,
                 },
                 "stderr_stream": {
                     "class": "FileStream",
                     "filename": u"%s/%s-stderr.log" % (logpath, name),
+                    "max_bytes": 2097152,
+                    # Removing backup logs for now to reduce clutter
+                    # "backup_count": 1,
                 }
 
             }